+++ bld/f77/wfc/c/symtab.c
@@ -2,7 +2,7 @@
 *
 *                            Open Watcom Project
 *
-* Copyright (c) 2002-2020 The Open Watcom Contributors. All Rights Reserved.
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -67,8 +67,8 @@ void    STInit( void ) {
 }
 
 
-sym_id SymLookup( const char *name, uint length )
-//===============================================
+sym_id SymLookup( const char *name, size_t length )
+//=================================================
 // Lookup a symbol in the symbol table.
 {
     sym_id    sym;
@@ -131,16 +131,16 @@ sym_id  LkBlkData( void ) {
 }
 
 
-sym_id SymFind( const char *name, uint length )
-//=============================================
+sym_id SymFind( const char *name, size_t length )
+//===============================================
 // Determine if a symbol exists in the symbol table.
 {
     return( STNameSearch( name, length ) );
 }
 
 
-static sym_id ComLookup( const char *name, uint length )
-//======================================================
+static sym_id ComLookup( const char *name, size_t length )
+//========================================================
 // Lookup a common block name.
 {
     sym_id      sym;
